extract BoxSelector, move ErgoBoxAssets to ergo_box;
