+++ sigma-tree/src/wallet.rs
@@ -1,5 +1,6 @@
 //! Wallet-related features for Ergo
 
+pub mod box_selector;
 pub mod secret_key;
 pub mod signing;
 pub mod tx_builder;
